refactor server so that we can test the socket logic.
